Spec cleanups (#103)
* spec: fix runtime deps

The binaries are in the greenboot package, not greenboot-rs
so move the runtime deps to the right location and remove
the obsolete deps from the root package.

Signed-off-by: Peter Robinson <pbrobinson@gmail.com>

* spec: Update spec description

Update the package description to something move descriptive.

Signed-off-by: Peter Robinson <pbrobinson@gmail.com>

* spec: cleanup spec formatting

The spec layout/formatting was a complete dog's breakfast
so lets clean it up so it's all uniform.

Signed-off-by: Peter Robinson <pbrobinson@gmail.com>

* spec: Update package name for systemd services

Fix the package name of the systemd services so they
actually get run.

Signed-off-by: Peter Robinson <pbrobinson@gmail.com>

* spec: drop unused conditionals, bump release

Drop global conditionals we don't use, bump the release.

Signed-off-by: Peter Robinson <pbrobinson@gmail.com>

* spec: Fix source URL

Fix the source URL so the file has identifiable naming.

Signed-off-by: Peter Robinson <pbrobinson@gmail.com>

* spec: fix up licensing

Because rust binaries are statically linked the whole binary
has to take into account the combined licenses of all the
components that go into making that binary. Add the cargo
licensing bits to the package and update the combined license
to reflect the output. The healthchecks are just BSD-3-Clause
so reflect that in the sub-package.

Signed-off-by: Peter Robinson <pbrobinson@gmail.com>

---------

Signed-off-by: Peter Robinson <pbrobinson@gmail.com>